Implemented Auhtorization for Update and Delete Methods in LisitingController

+++ api/V1/controllers/ListingController.php
@@ -7,24 +7,24 @@ use Core\DB;
 use PDO;
 use Firebase\JWT\Key;
 use Firebase\JWT\JWT;
-
+use Valitron\Validator;
+use API\V1\Controllers\AuthController;
 
 
 class ListingController
 { 
     
-    private $con, $db, $key, $auth;
+    private $con, $db, $key, $auth_user_id;
 
     public function __construct()
     {
         $this->db = new DB();
         $this->con = $this->db->connect("real-estate", "root", "");
-        $this->auth = new AuthController();
+        $this->auth_user_id = AuthController::get_user();
     }
 
     public function index(ServerRequest $request)
     {
-
             //IMPROVEMENTS: 
             //Make response HATEOS compliant
             $statement = $this->con->prepare("SELECT * FROM `listings`");
@@ -41,49 +41,78 @@ class ListingController
         $statement->bindParam(":id", $id);
         $statement->execute();
 
+        if($statement->rowCount() <= 0)
+        {
+            return new JsonResponse(["error" => "Listing not found"], 404);
+        }
         return new JsonResponse($statement->fetch(PDO::FETCH_ASSOC), 200);
     }
 
     public function create(ServerRequest $request)
     {
-        //IMPROVEMENTS:
-        //Request should be validated before inserting (compulsory fields should be enforced)
-        //Add $allowedFields variable for validation
-
         //get request content
         $data = $request->getParsedBody();
-        try 
+        $data["user_id"] = AuthController::get_user();
+
+        
+        $allowedFields = ["user_id", "name", "description", "pictures", "details", "location"];
+        
+        //validate
+        $validator = new Validator($data);
+        $validator->rule('required',  $allowedFields);
+
+        if($validator->validate())
         {
-            $statement = $this->con->prepare("INSERT INTO `listings` (`name`, `description`, `pictures`, `details`, `location`) 
-            VALUES (:name, :description, :pictures, :details, :location)"); 
+            try 
+            {
+                $statement = $this->con->prepare("INSERT INTO `listings` (`user_id`, `name`, `description`, `pictures`, `details`, `location`) 
+                VALUES (:user_id, :name, :description, :pictures, :details, :location)"); 
 
-            $statement->bindParam(":name", $data['name']);
-            $statement->bindParam(":description", $data['description']);
-            $statement->bindParam(":pictures", $data['pictures']);
-            $statement->bindParam(":details", $data['details']);
-            $statement->bindParam(":location", $data['location']);
+                $this->db->bindAllParams($statement, $data, $allowedFields);
+    
+                $statement->execute();
+    
+                $id = $this->con->lastInsertId();
+    
+            } catch (Exception $e)
+            {
+                return json_encode(["error" => $e]);
+            }
+          
+           return $this->show($id);
 
-            $statement->execute();
-
-            $id = $this->con->lastInsertId();
-
-        } catch (Exception $e)
+        } else
         {
-            return json_encode(["error" => $e]);
+            return new JsonResponse(["errors" => $validator->errors()], 504);
         }
-      
-       return $this->show($id);
+
     }
 
     public function update(ServerRequest $request, $id)
     {
-        //IMPROVEMENTS: 
-        //You should only be able to update your own listing
-        //Add user id to `listing` table
-        //reject empty request
 
+        //reject empty request
         $data =  $request->getQueryParams();
-        $allowedFields = [ "name", "description", "pictures", "details","California","is_available"];
+        if (is_null($data) || empty($data))
+        {
+            return new JsonResponse(["error" => "No value to update"], 401);
+        }
+
+        //Check if listing exists
+        $listing = $this->show($id);
+        if($listing->getStatusCode() == 404)
+        {
+            return new JsonResponse(["error" => "Listing does not exist"], 404);
+        }
+
+        //You should only be able to update your own listing
+        $user_id = json_decode($listing->getBody()->getContents())->user_id;
+        if($user_id != $this->auth_user_id)
+        {
+            return new JsonResponse(["error" => "You are not authorized to carry out this operation"], 401);
+        }
+
+        $allowedFields = ["name", "description", "pictures", "details","California", "is_available"];
 
         //get query parameter
         $queryParams = $this->db->getQueryParams($data, $allowedFields);
@@ -99,8 +128,19 @@ class ListingController
 
     public function destroy($id)
     {
-         //IMPROVEMENTS: 
-        //You should only be able to delete your own listing
+        //Check if listing exists
+        $listing = $this->show($id);
+        if($listing->getStatusCode() == 404)
+        {
+            return new JsonResponse(["error" => "Listing does not exist"], 404);
+        }
+
+        //You should only be able to update your own listing
+        $user_id = json_decode($listing->getBody()->getContents())->user_id;
+        if($user_id != $this->auth_user_id)
+        {
+            return new JsonResponse(["error" => "You are not authorized to carry out this operation"], 401);
+        }
 
         $statement = $this->con->prepare("DELETE FROM `listings` WHERE id=:id");
         $statement->bindParam(":id", $id);
